+++ spring-extension-context/src/test/java/com/livk/context/mybatis/type/postgresql/PostgresJsonTypeHandlerTest.java
@@ -16,12 +16,33 @@
 
 package com.livk.context.mybatis.type.postgresql;
 
+import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import com.livk.commons.jackson.util.JsonMapperUtils;
 import com.livk.testcontainers.containers.PostgresqlContainer;
+import com.zaxxer.hikari.HikariDataSource;
+import lombok.Data;
+import org.apache.ibatis.annotations.Insert;
+import org.apache.ibatis.annotations.Mapper;
+import org.apache.ibatis.annotations.Select;
+import org.apache.ibatis.logging.stdout.StdOutImpl;
+import org.apache.ibatis.mapping.Environment;
+import org.apache.ibatis.session.SqlSessionFactory;
+import org.apache.ibatis.session.defaults.DefaultSqlSessionFactory;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.mybatis.spring.SqlSessionTemplate;
+import org.mybatis.spring.mapper.MapperScannerConfigurer;
+import org.mybatis.spring.transaction.SpringManagedTransactionFactory;
+import org.postgresql.Driver;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.testcontainers.properties.TestcontainersPropertySourceAutoConfiguration;
 import org.springframework.boot.testcontainers.service.connection.ServiceConnection;
+import org.springframework.boot.testcontainers.service.connection.ServiceConnectionAutoConfiguration;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
 import org.springframework.core.io.ClassPathResource;
 import org.springframework.jdbc.datasource.init.ScriptUtils;
 import org.springframework.test.context.DynamicPropertyRegistry;
@@ -32,6 +53,7 @@ import org.testcontainers.junit.jupiter.Testcontainers;
 
 import javax.sql.DataSource;
 import java.sql.SQLException;
+import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
@@ -42,7 +64,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
  *
  * @author livk
  */
-@SpringJUnitConfig(MybatisConfig.class)
+@SpringJUnitConfig(PostgresJsonTypeHandlerTest.MybatisConfig.class)
 @Testcontainers(disabledWithoutDocker = true, parallel = true)
 class PostgresJsonTypeHandlerTest {
 
@@ -87,6 +109,75 @@ class PostgresJsonTypeHandlerTest {
 		assertEquals(user.getPassword(), first.getPassword());
 		assertEquals(user.getDes(), first.getDes());
 		assertEquals("livk", first.getDes().get("mark").asText());
+	}
+
+	@Configuration
+	@Import({ ServiceConnectionAutoConfiguration.class, TestcontainersPropertySourceAutoConfiguration.class })
+	static class MybatisConfig {
+
+		@Bean(destroyMethod = "close")
+		public HikariDataSource dataSource(@Value("${spring.datasource.url}") String url,
+				@Value("${spring.datasource.username}") String username,
+				@Value("${spring.datasource.password}") String password) {
+			HikariDataSource dataSource = new HikariDataSource();
+			dataSource.setDriverClassName(Driver.class.getName());
+			dataSource.setJdbcUrl(url);
+			dataSource.setUsername(username);
+			dataSource.setPassword(password);
+			return dataSource;
+		}
+
+		@Bean
+		public SqlSessionFactory sqlSessionFactory(DataSource dataSource) {
+			Environment environment = new Environment("mybatis", new SpringManagedTransactionFactory(), dataSource);
+			org.apache.ibatis.session.Configuration configuration = new org.apache.ibatis.session.Configuration(
+					environment);
+			configuration.setMapUnderscoreToCamelCase(true);
+			configuration.setLogImpl(StdOutImpl.class);
+			PostgresJsonTypeHandler typeHandler = new PostgresJsonTypeHandler(new ObjectMapper());
+			configuration.getTypeHandlerRegistry().register(typeHandler);
+			return new DefaultSqlSessionFactory(configuration);
+		}
+
+		@Bean
+		public SqlSessionTemplate sqlSessionTemplate(SqlSessionFactory sqlSessionFactory) {
+			return new SqlSessionTemplate(sqlSessionFactory);
+		}
+
+		@Bean
+		public MapperScannerConfigurer mapperScannerConfigurer() {
+			MapperScannerConfigurer configurer = new MapperScannerConfigurer();
+			configurer.setSqlSessionFactoryBeanName("sqlSessionFactory");
+			configurer.setSqlSessionTemplateBeanName("sqlSessionTemplate");
+			configurer.setBasePackage(this.getClass().getPackageName());
+			configurer.setAnnotationClass(Mapper.class);
+			return configurer;
+		}
+
+	}
+
+	@Mapper
+	public interface UserMapper {
+
+		@Insert("insert into \"user\" (username,password,des) values (#{username},#{password},#{des})")
+		int insert(User user);
+
+		@Select("select * from \"user\"")
+		List<User> selectList();
+
+	}
+
+	@Data
+	static public class User {
+
+		private Long id;
+
+		private String username;
+
+		private String password;
+
+		private JsonNode des;
+
 	}
 
 }
